Start tests for interface between runtime and host
